requirements.txt
aiomysql==0.2.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
orjson==3.8.10
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.0.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
